Rename Transaction to Entry
Transaction will inevitably cause problems/confusion with database
transactions. It is also not quite the correct domain term as entries
need not correspond one-to-one with uploaded transactions: could
split a transaction into several entries, for example.
Transaction should still be used for the upload though: as at this
stage they are transactions in the banking sense, and each transaction
is added as an entry.