Remove unnecessary failing view specs
Date format can be tested in isolation, and we trust that the view
uses it correctly. Money format is untested as it is built-in.
The other tests were checking that the view does what the view is
fundamentally required to do - these are fragile and unnecesssary.
A change in the view is a deliberate change in the view and it
can only affect the view itself, nothing depends on the view.
Kept 1 test for the conditional - this logic belongs here rather
than in the controller, and it also appears to be faster to test.
The other side of the conditional is core functionality and will be
touched by many acceptance tests, but I think I'd notice if I broke
the home page...