Fix N+1 query bug
For some reason accessing tag names via tag_names causes the queries
to be carried out even when tags are eager-loaded.
This also means we are eager loading descriptions (and anything else
we add to tags) when all we need is the names.
HomeController specs cleaned up as 3 tests were essentially testing
the same thing: result of query is assigned