Bind event to document to avoid Turbolinks issues
The event was binding to the links on page load, meaning that when the
page changed (without calling a page load event) the bindings were
lost.

Fix is to use event delegation to bind directly to document and have
this find the relevant element on-the-fly.
The cancel button can safely be bound directly when the form is shown.